indexer-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
pydantic-settings
bcrypt
azure-identity
azure-keyvault-secrets
python-jose[cryptography]
python-dotenv
